// src/modules/car/types.ts
export enum CarEnumTypes {
    GET_USER_CAR_REQUEST = 'GET_USER_CAR_REQUEST',
    GET_USER_CAR_RESPONSE = 'GET_USER_CAR_RESPONSE',
    GET_USER_CAR_ERROR = 'GET_USER_CAR_ERROR',

    POST_USER_CAR_REQUEST = 'POST_USER_CAR_REQUEST',
    POST_USER_CAR_RESPONSE = 'POST_USER_CAR_RESPONSE',
    POST_USER_CAR_ERROR = 'POST_USER_CAR_ERROR',

    PUT_USER_CAR_REQUEST = 'PUT_USER_CAR_REQUEST',
    PUT_USER_CAR_RESPONSE = 'PUT_USER_CAR_RESPONSE',
    PUT_USER_CAR_ERROR = 'PUT_USER_CAR_ERROR',
}

export interface PostCar {
    title: string
    number: string
}